apps/worker/requirements.txt
playwright==1.58.0
redis==5.0.8
sqlalchemy==2.0.36
psycopg[binary]==3.3.3
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-slugify==8.0.4
